Fixed Two Bugs found during Scarlett Development
If you configure default options on an Endpoint, they wouldn’t stick,
due to a blank $params array being passed to Endpoint on Instantiation.

If you configured a default set of Data on an Endpoint, an error would
be thrown that data was missing even though the defaults covered all
data.