Minor Change to SetCredentials on SugarClient Abstract
When working with Scarlett, and SI objects I came to the point where I
wanted the same functionality that was in the basic SugarAPI object,
that comes with the composer package, in that I wanted to just pass
Username and Password to the client object, and not Client
ID/Secret/Platform as well. So I moved the setCredentials method from
Basic SugarAPI object, to the Abstract so that Credentials are now
appended to, rather than completely written to the property. I did
however not set defaults on the values of the credential properties, as
I have on the base SugarAPI Object, since I would prefer people
configure Client and Client Secrets rather than always use the base one
that comes with Sugar.