Added in API Versioning
To accommodate for the API Versioning that is possible with
customizations, I added in an APIVersion and SetVersion/GetVersion
method on the SugarClient. This can be passed to ConfigureAPI Url in
Helpers as well, to allow for use cases where people need to use
multiple Endpoints that are the same, but are different versions.
Change Configure Api Url to use Regular Expression checks, for
simplication of code.