Check for Content-Disposition Header
When getting a Filename from the Headers, I need to make sure I’m
looking at the Content-Disposition Header. I came across a scenario
where after a 302 Redirect, the headers contain Filename twice, which
caused the Filename to be incorrect and would crash the application
since fopen couldn’t be called on the ridiculous filename.