Bulk API + Filter Changes + PHPDoc Updates + Unit Tests Started
Added Bulk API Endpoint, with custom BulkRequest Data Object to parse
Endpoints and Request Objects, so that you don’t have to manually build
out the Bulk Request yourself.

Added an SugarEndpointInterface to allow for a “compileRequest” method,
which allows for easier debugging of Endpoints, and helps the Bulk API
Endpoint.

Cleaned up PHP Inline Docs, and made unit tests for everything except
the Endpoints. Endpoint Unit Tests are started, but a lot more work to
be done.