Count Endpoints + Filter Changes
FilterData object is better coupled with Filter parameter on the
Endpoint objects. If you set a default filter on the Endpoint via Data
Properties, it will push to the Filter Object on initialize, and when
you build on top of the Filter object it merges with the default
applied. It also will reset the parameter when it gets reset on the
Endpoint.

Added count endpoints to FilterRelated/GetRelated and ModuleFilter
Endpoints.